Some fixes

+++ parallel_code/rate4site.py
@@ -7,8 +7,11 @@ def get_rate4site(MSA_path, tree_path, output_path):
     rate4site_command = "{rate4site_exe_path}  -bn -s {msa_path} -t {tree_path} -o {output_path}".format(
         rate4site_exe_path = RATE4SITE_COMMAND_PREFIX, msa_path = MSA_path, tree_path = tree_path, output_path = output_path)
     subprocess.run(rate4site_command, shell=True)
-    os.remove("r4s.res")
-    os.remove("r4sOrig.res")
+    if os.path.exists("r4s.res"):
+        os.remove("r4s.res")
+    if os.path.exists("r4sOrig.res"):
+        os.remove("r4sOrig.res")
+    return rate4site_command
 
 
 def parse_rate4site(rate4site_output_path):
@@ -28,6 +31,5 @@ def parse_rate4site(rate4site_output_path):
 # scores = parse_rate4site(output_path)
 # print(scores)
 # print(len(scores))
-# scores = parse_rate4site("r4sOrig.res")
 # print(scores)
 # print(len(scores))
